refactor importer into library
