+++ crates/importer/Cargo.toml
@@ -3,6 +3,9 @@ name = "importer"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+path = "src/lib.rs"
+
 [[bin]]
 name = "chatgpt-importer"
 path = "src/main.rs"
